Split runner managed temporary project tests
